Rename to tree_concern
Rename to tree_concern to match gem naming conventions.

Tree moved to TreeConcern::Tree
TreeMigration moved to TreeConcern::Migration
