--- lib/tree-migration.rb
@@ -1,7 +1,0 @@
-module TreeMigration
-  extend ActiveSupport::Concern
-  
-  def add_tree table_name
-      add_reference table_name, :parent, foreign_key: {to_table: table_name}
-  end
-end
