+++ lib/tree_concern.rb
@@ -1,0 +1,6 @@
+require_relative 'tree_concern/tree'
+require_relative 'tree_concern/migration'
+
+module TreeConcern
+    
+end
